feat(selenium): add url assert functions to navigator class and refactor tests
