+++ selenium/src/keywords/navigator.py
@@ -26,15 +26,19 @@ class Navigator:
             Pages.REGISTER, self.driver
         )
 
-    def _expect(self, url: str, *, should_assert: bool = True, should_wait: bool = True):
+    def url_to_be(self, url: str, *, should_assert: bool = True, should_wait: bool = True):
         at_url: bool = self.custom_selenium.url_to_be(url, should_wait=should_wait)
         if should_assert:
             assert at_url, f"Current page is not at {url}"
+        return at_url
+
+    def url_contains(self, url: str, *, should_wait: bool = True):
+        return self.custom_selenium.url_contains(url, should_wait=should_wait)
 
     def navigate_to_login(self, *, should_assert: bool = True, should_wait: bool = True):
         login_url: str = self.login_actions.goto()
-        self._expect(login_url, should_assert=should_assert, should_wait=should_wait)
+        self.url_to_be(login_url, should_assert=should_assert, should_wait=should_wait)
 
     def navigate_to_register(self, *, should_assert: bool = True, should_wait: bool = True):
         register_url: str = self.register_actions.goto()
-        self._expect(register_url, should_assert=should_assert, should_wait=should_wait)
+        self.url_to_be(register_url, should_assert=should_assert, should_wait=should_wait)
