+++ selenium/tests/login/test_login.py
@@ -11,6 +11,8 @@ if typing.TYPE_CHECKING:
     from src.pom.query.login.login import LoginQuery
     from src.pom.query.user_home.user_home import UserHomeQuery
 
+from src.keywords.navigator import Navigator
+
 from selenium.webdriver.remote.webdriver import WebDriver
 
 
@@ -24,13 +26,15 @@ class TestLogin:
         _admin_home: AdminHomeQuery = factory.QueriesFactory.get(
             Pages.ADMIN_HOME, driver=driver  # type: ignore
         )
+        navigator = Navigator(driver=driver)
+
         _login_actions.fill_email(admin_account.email)
         _login_actions.fill_password(admin_account.password)
         _login_actions.click_login()
 
+        assert navigator.url_contains("admin")
         assert _admin_home.has_welcome_message(for_user=admin_account.name)
         assert _admin_home.is_store_description_visible()
-        assert "admin" in driver.current_url
 
     @pytest.mark.login_002
     def test_user_redirected_to_home(self, driver: WebDriver, user_account):
@@ -40,12 +44,13 @@ class TestLogin:
         _user_home: UserHomeQuery = factory.QueriesFactory.get(
             Pages.USER_HOME, driver=driver  # type: ignore
         )
+        navigator = Navigator(driver=driver)
 
         _login_actions.fill_email(user_account.email)
         _login_actions.fill_password(user_account.password)
         _login_actions.click_login()
 
-        assert "admin" not in driver.current_url
+        assert navigator.url_contains("admin") is False
         assert _user_home.is_store_visible()
 
     @pytest.mark.login_003
@@ -53,10 +58,11 @@ class TestLogin:
         _login_actions: LoginActions = factory.ActionsFactory.get(
             Pages.LOGIN, driver=driver  # type: ignore
         )
+        navigator = Navigator(driver=driver)
 
         _login_actions.click_register()
 
-        assert driver.current_url == f"{URL.base_url}{URL.register}"
+        assert navigator.url_to_be(f"{URL.base_url}{URL.register}", should_assert=False)
 
     @pytest.mark.login_004
     def test_alert_email_is_required(self, driver: WebDriver):
